Properties to limit maximum of export results.
This is for csv,warc and warc-expanded

These are the properties and default values if no defined.

#Limit export size
#10M for CSV , 1M warc and csv, 10K for warc-expanded
#For warc.expanded the total number of documents can be many times the
max-results size.
export.csv.maxresults=10000000
export.warc.maxresults=1000000
export.warc.expanded.maxresults=10000

+++ src/main/java/dk/kb/netarchivesuite/solrwayback/export/StreamingSolrExportBufferedInputStream.java
@@ -11,8 +11,8 @@ public class StreamingSolrExportBufferedInputStream extends InputStream {
 
   private int index;
   private List<byte[]> inputBuffer = new ArrayList<>();
-  private final int maxLines;
-  private int linesRead;
+  private final long maxLines;
+  private long linesRead;
   private  SolrStreamingLineBasedExportClientInterface solrClient;
   private final int solrPagingBufferSize;
 
@@ -47,7 +47,7 @@ public class StreamingSolrExportBufferedInputStream extends InputStream {
     return result;
   }
 
-  public StreamingSolrExportBufferedInputStream(SolrStreamingLineBasedExportClientInterface solrClient, int maxLines) {
+  public StreamingSolrExportBufferedInputStream(SolrStreamingLineBasedExportClientInterface solrClient, long maxLines) {
     this.solrClient = solrClient;
     this.maxLines = maxLines;
     this.solrPagingBufferSize = solrClient.getPageSize();
